# app/Http/Controllers/TaskController.php
<?php

namespace App\Http\Controllers;

use App\Models\Task;
use Illuminate\Http\Request;
use Illuminate\Support\Facades\Validator;


/**
 * @OA\Info(
 *     title="Tasks API",
 *     version="1.0.0",
 *     description="API for managing tasks",
 *     @OA\Contact(
 *         email="toni.fernandez@cirvianum.cat"
 *     )
 * )
 *
 * * @OA\Server(
 *     url="http://localhost",
 *     description="Local development server"
 * )
 */


class TaskController extends Controller
{
    /**
     * Display a listing of the resource.
     */

    /**
     * Tasks List
     * @OA\Get (
     *     path="/api/tasks",
     *     tags={"get_tasks"},
     *     @OA\Response(
     *         response=200,
     *         description="OK",
     *         @OA\JsonContent(
     *             @OA\Property(
     *                 type="array",
     *                 property="rows",
     *                 @OA\Items(
     *                     type="object",
     *                     @OA\Property(
     *                         property="id",
     *                         type="number",
     *                         example="1"
     *                     ),
     *                     @OA\Property(
     *                         property="name",
     *                         type="string",
     *                         example="name task"
     *                     ),
     *                     @OA\Property(
     *                         property="descriptioon",
     *                         type="string",
     *                         example="Non placeat illum ex dolorem sint fugit natus."
     *                     ),
     *                      @OA\Property(
     *                         property="user_id",
     *                         type="number",
     *                         example="1"
     *                     ),
     *                     @OA\Property(
     *                         property="created_at",
     *                         type="string",
     *                         example="2023-02-23T00:09:16.000000Z"
     *                     ),
     *                     @OA\Property(
     *                         property="updated_at",
     *                         type="string",
     *                         example="2023-02-23T12:33:45.000000Z"
     *                     )
     *                 )
     *             )
     *         )
     *     )
     * )
     */

    public function index()
    {
        try {
            $tasks = Task::all();
            return response()->json([
                'status' => true,
                'tasks' => $tasks,
                'httpCode' => 200
            ]);
        } catch (\Throwable $th) {
            return response()->json([
                'status' => false,
                'message' => $th->getMessage(),
                'httpCode' => 500
            ]);
        }
    }

    /**
     * Show the form for creating a new resource.
     */
    public function create()
    {
        //
    }

    /**
     * Store a newly created resource in storage.
     */

    /**
     * Crea uan tasca
     * @OA\Post (
     *     path="/api/tasks",
     *     tags={"create_task"},
     *     @OA\Response(
     *         response=201,
     *         description="Task created successfully",
     *     ),
     *    @OA\RequestBody(
     *         required=true,
     *         description="User details",
     *         @OA\JsonContent(
     *             @OA\Property(property="name", type="string", example="Task swaggger"),
     *             @OA\Property(property="description", type="string", example="This a typical descrition for a task"),
     *             @OA\Property(property="user_id", type="string", example="2"),
     *
     *
     *         )
     *     ),
     *      @OA\Response(
     *          response=422,
     *          description="Validation error",
     *      ),
     *      @OA\Response(
     *          response=405,
     *          description="Validation error",
     *      )
     *)
     *)
     */

    public function store(Request $request)
    {
        $validator = Validator::make($request->all(), [
            'name' => ['required', 'string', 'max:50'],
            'description' => ['required', 'string'],
            'user_id' => ['required', 'exists:users,id']
        ]);

        if ($validator->fails()) {
            return response()->json([
                'status' => false,
                'message' => $validator->errors(),
                'httpCode' => 422
            ]);
        }

        try {
            $task = Task::create($request->all());
            return response()->json([
                'status' => true,
                'message' => 'Task afegida a la base de dades',
                'httpCode' => 201
            ]);
        } catch (\Throwable $th) {
            return response()->json([
                'status' => false,
                'message' => $th->getMessage(),
                'httpCode' => 500
            ]);
        }
    }

    /**
     * Display the specified resource.
     */
    /**
     * @OA\Get(
     *     path="/api/tasks/{id}",
     *     summary="Get task by id",
     *     description="Return a task by Id",
     *     operationId="getTaskById",
     *     tags={"get_task_bt_id"},
     *
     *     @OA\Parameter(
     *         name="id",
     *         in="path",
     *         description="ID de la tarea a recuperar",
     *         required=true,
     *         @OA\Schema(type="string")
     *     ),
     *
     *     @OA\Response(
     *         response=200,
     *         description="Tarea encontrada correctamente",
     *         @OA\JsonContent(
     *             @OA\Property(property="status", type="boolean", example=true),
     *             @OA\Property(property="task", type="object",
     *                 @OA\Property(property="id", type="string", example="1"),
     *                 @OA\Property(property="title", type="string", example="Completar el informe"),
     *                 @OA\Property(property="description", type="string", example="El informe debe ser entregado el lunes"),
     *                 @OA\Property(property="status", type="string", example="pending")
     *             ),
     *             @OA\Property(property="httpCode", type="integer", example=200)
     *         )
     *     ),
     *
     *     @OA\Response(
     *         response=404,
     *         description="Tarea no encontrada",
     *         @OA\JsonContent(
     *             @OA\Property(property="status", type="boolean", example=false),
     *             @OA\Property(property="message", type="string", example="Task not found"),
     *             @OA\Property(property="httpCode", type="integer", example=404)
     *         )
     *     ),
     *
     *     @OA\Response(
     *         response=500,
     *         description="Error interno del servidor",
     *         @OA\JsonContent(
     *             @OA\Property(property="status", type="boolean", example=false),
     *             @OA\Property(property="message", type="string", example="Error inesperado"),
     *             @OA\Property(property="httpCode", type="integer", example=500)
     *         )
     *     )
     * )
     */
    public function show(string $id)
    {
        $task = Task::find($id);
        if (is_null($task)) {
            return response()->json([
                'status' => false,
                'message' => 'Task not found',
                'httpCode' => 404
            ]);
        }

        try {
            return response()->json([
                'status' => true,
                'task' => $task,
                'httpCode' => 404
            ]);
        } catch (\Throwable $th) {
            return response()->json([
                'status' => false,
                'message' => $th->getMessage(),
                'httpCode' => 500
            ]);
        }
    }

    /**
     * Show the form for editing the specified resource.
     */
    public function edit(string $id)
    {
        //
    }

    /**
     * Update the specified resource in storage.
     */

    /**
     * @OA\Put(
     *     path="/api/tasks/{id}",
     *     summary="Udpate task by id",
     *     description="Update a task by Id",
     *     operationId="updateTask",
     *     tags={"update_task"},
     *
     *     @OA\Parameter(
     *         name="id",
     *         in="path",
     *         description="ID de la tarea a actualizar",
     *         required=true,
     *         @OA\Schema(type="string")
     *     ),
     *
     *     @OA\RequestBody(
     *         required=true,
     *         description="Datos necesarios para actualizar la tarea",
     *         @OA\JsonContent(
     *             required={"name", "description", "user_id"},
     *             @OA\Property(property="name", type="string", example="Nueva tarea actualizada", description="Nombre de la tarea"),
     *             @OA\Property(property="description", type="string", example="Descripción actualizada de la tarea", description="Descripción de la tarea"),
     *             @OA\Property(property="user_id", type="integer", example=1, description="ID del usuario responsable")
     *         )
     *     ),
     *
     *     @OA\Response(
     *         response=200,
     *         description="Tarea actualizada correctamente",
     *         @OA\JsonContent(
     *             @OA\Property(property="status", type="boolean", example=true),
     *             @OA\Property(property="message", type="string", example="Task updated"),
     *             @OA\Property(property="httpCode", type="integer", example=200)
     *         )
     *     ),
     *
     *     @OA\Response(
     *         response=404,
     *         description="Tarea no encontrada",
     *         @OA\JsonContent(
     *             @OA\Property(property="status", type="boolean", example=false),
     *             @OA\Property(property="message", type="string", example="No query results for model [Task]"),
     *             @OA\Property(property="httpCode", type="integer", example=404)
     *         )
     *     ),
     *
     *     @OA\Response(
     *         response=422,
     *         description="Error de validación",
     *         @OA\JsonContent(
     *             @OA\Property(property="status", type="boolean", example=false),
     *             @OA\Property(property="message", type="object",
     *                 @OA\Property(property="name", type="array",
     *                     @OA\Items(type="string", example="The name field is required.")
     *                 ),
     *                 @OA\Property(property="description", type="array",
     *                     @OA\Items(type="string", example="The description field is required.")
     *                 ),
     *                 @OA\Property(property="user_id", type="array",
     *                     @OA\Items(type="string", example="The user_id field must exist.")
     *                 )
     *             ),
     *             @OA\Property(property="httpCode", type="integer", example=422)
     *         )
     *     ),
     *
     *     @OA\Response(
     *         response=500,
     *         description="Error interno del servidor",
     *         @OA\JsonContent(
     *             @OA\Property(property="status", type="boolean", example=false),
     *             @OA\Property(property="message", type="string", example="Error inesperado"),
     *             @OA\Property(property="httpCode", type="integer", example=500)
     *         )
     *     )
     * )
     */
    public function update(Request $request, string $id)
    {
        try {
            $task = Task::findOrFail($id);
        } catch (\Throwable $th) {
            return response()->json([
                'status' => false,
                'message' => $th->getMessage(),
                'httpCode' => 404
            ]);
        }

        $validator = Validator::make($request->all(), [
            'name' => ['required', 'string', 'max:50'],
            'description' => ['required', 'string'],
            'user_id' => ['required', 'exists:users,id']
        ]);

        if ($validator->fails()) {
            return response()->json([
                'status' => false,
                'message' => $validator->errors(),
                'httpCode' => 422
            ]);
        }

        try {
            $task->update($request->all());
            return response()->json([
                'status' => true,
                'message' => 'Task updated',
                'httpCode' => 200
            ]);
        } catch (\Throwable $th) {
            return response()->json([
                'status' => false,
                'message' => $th->getMessage(),
                'httpCode' => 404
            ]);
        }
    }

    /**
     * Remove the specified resource from storage.
     */

    /**
     * @OA\Delete(
     *     path="/api/tasks/{id}",
     *     summary="Eliminar una tarea",
     *     description="Elimina una tarea específica por su ID. Devuelve un error si la tarea no existe.",
     *     operationId="deleteTask",
     *     tags={"delete_task"},
     *
     *     @OA\Parameter(
     *         name="id",
     *         in="path",
     *         description="ID de la tarea a eliminar",
     *         required=true,
     *         @OA\Schema(type="string")
     *     ),
     *
     *     @OA\Response(
     *         response=200,
     *         description="Tarea eliminada correctamente",
     *         @OA\JsonContent(
     *             @OA\Property(property="status", type="boolean", example=true),
     *             @OA\Property(property="message", type="string", example="Task deleted successfully"),
     *             @OA\Property(property="httpCode", type="integer", example=200)
     *         )
     *     ),
     *
     *     @OA\Response(
     *         response=404,
     *         description="Tarea no encontrada",
     *         @OA\JsonContent(
     *             @OA\Property(property="status", type="boolean", example=false),
     *             @OA\Property(property="message", type="string", example="No query results for model [Task]"),
     *             @OA\Property(property="errorCode", type="integer", example=404)
     *         )
     *     ),
     *
     *     @OA\Response(
     *         response=500,
     *         description="Error interno del servidor",
     *         @OA\JsonContent(
     *             @OA\Property(property="status", type="boolean", example=false),
     *             @OA\Property(property="message", type="string", example="Error inesperado al eliminar la tarea"),
     *             @OA\Property(property="httpCode", type="integer", example=500)
     *         )
     *     )
     * )
     */

    public function destroy(string $id)
    {
        try {
            Task::findOrFail($id);
        } catch (\Exception $e) {
            return response()->json([
                'status' => false,
                'message' => $e->getMessage(),
                'errorCode' => 404
            ]);
        }

        try {
            Task::destroy($id);
            return response()->json([
                'status' => true,
                'message' => 'Task deleted successfully',
                'httpCode' => 200
            ]);
        } catch (\Throwable $th) {
            return response()->json([
                'status' => false,
                'message' => $th->getMessage(),
                'httpCode' => 500
            ]);
        }
    }
}