# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tasks API",
        "description": "API for managing tasks",
        "contact": {
            "email": "toni.fernandez@cirvianum.cat"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Local development server"
        }
    ],
    "paths": {
        "/api/tasks": {
            "get": {
                "tags": [
                    "get_tasks"
                ],
                "summary": "Tasks List",
                "operationId": "471d938a3bfaeb6fbd981d7c3e2728b0",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "rows": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "name task"
                                                    },
                                                    "descriptioon": {
                                                        "type": "string",
                                                        "example": "Non placeat illum ex dolorem sint fugit natus."
                                                    },
                                                    "user_id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2023-02-23T00:09:16.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2023-02-23T12:33:45.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "create_task"
                ],
                "summary": "Crea uan tasca",
                "operationId": "f803ebdb2300acdb7e93bbe8d0b9a2e8",
                "requestBody": {
                    "description": "User details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Task swaggger"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "This a typical descrition for a task"
                                    },
                                    "user_id": {
                                        "type": "string",
                                        "example": "2"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Task created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "405": {
                        "description": "Validation error"
                    }
                }
            }
        },
...
    },
...
}